Updated hidden layer width data type to int16.

+++ ann/circular_roa/main.py
@@ -195,7 +195,7 @@ problem_specifications.save( save_path, r'problem_specifications.pkl' )
 # Store the network parameters.
 activation_function = 'sigmoid'                                                                # [-] Activation function (e.g., tanh, sigmoid, etc.)
 num_hidden_layers = torch.tensor( 5, dtype = torch.uint8, device = device )                 # [#] Number of hidden layers.
-hidden_layer_widths = torch.tensor( 500, dtype = torch.uint8, device = device )              # [#] Hidden layer widths.
+hidden_layer_widths = torch.tensor( 500, dtype = torch.int16, device = device )              # [#] Hidden layer widths.
 
 # This set works for variational loss integration order 1.
 num_training_data = torch.tensor( int( 100e3 ), dtype = torch.int32, device = device )      # [#] Number of training data points.
